Add examples and fix bug in common.h

+++ lib/common.cpp
@@ -6,6 +6,6 @@ int cmpint(const void *a, const void *b){
 	return greater(a, b, int);
 }
 
-double cmpdouble(const void *a, const void *b){
+int cmpdouble(const void *a, const void *b){
 	return greater(a, b, double);
 }
